refactor(core/mercury): unify confirm TOS
[no changelog]
